requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
fastapi
deepeval

-e .
